byond-tracy forces a hard reboot (#90432)
byond-tracy lacks the ability to clean up its hooks, so it will error if
it tries to initialize again after a normal reboot.

This forces a hard reboot (regardless of the `rounds_until_hard_restart`
config) if byond-tracy is enabled.

I also folded the hard reboot checks into a new proc,
`/world/proc/check_hard_reboot()`, for the sake of cleanliness, and
added a `flush()` function to `/datum/tracy`, although currently it's
not called (it also gets flushed by the `destroy` call) - this is just
for the sake of documenting that it exists in case anyone wants to use
it in the future, altho if anyone knows good places to call it, let me
know.

Less likely to cause problems.

No player-facing changes.